Create Codes
Task Description 
 
First, I wrote a function for translating DNA to protein. This function simulates and generates a logistic population growth curve. The function includes 2 extra parameters that randomize the length of the lag phase and the exponential phase. The growth curve displays: Population Size vs Time.

 The function also generate a dataframe with 100 different growth curves.
Next another function was generated to determine the time to reach 80% of the maximum growth; usually the carrying capacity.

Finally, the hamming distance between my Slack username and twitter/X handle (which was padded with other letters) was determined using another function created.